Separate graphic code from main
